requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
audioread==3.0.1

# Machine Learning
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3
scipy==1.11.3

# Utilities
python-dotenv==1.0.0